Moved functions for default parser rules to new namespace
